Build springboot from HEAD
Previously, the springboot formula would only download and install the
latest milestone.  For people testing snapshots (and not wanting to do
a manual installation), this wasn't entirely helpful.  This change
adds support for Homebrew's unstable version support[1].  Users can
now do a

    brew install springboot --HEAD

and get an install built from the HEAD of the Spring Boot repo.

[1]: https://github.com/Homebrew/homebrew/wiki/Formula-Cookbook#unstable-versions-head-devel